+++ CHANGELOG.md
@@ -7,3 +7,4 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - sends KeepAlive pings until report is sent
+- Adds zipper module for zipping reports directory